chore(suite-desktop): remove suite imports
