+++ packages/suite-build/configs/web.webpack.config.ts
@@ -4,8 +4,7 @@ import CopyPlugin from 'copy-webpack-plugin';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
 
-import routes from '../../suite/src/config/suite/routes';
-import { FLAGS } from '@suite-common/suite-config';
+import { FLAGS, routes } from '@suite-common/suite-config';
 
 import { assetPrefix, isDev } from '../utils/env';
 import { getPathForProject } from '../utils/path';
