+++ packages/suite/src/actions/wallet/selectedAccountActions.ts
@@ -2,11 +2,11 @@ import { accountsActions, blockchainActions, discoveryActions } from '@suite-com
 import { getAccountNetwork } from '@suite-common/wallet-utils';
 import { SelectedAccountStatus } from '@suite-common/wallet-types';
 import { DiscoveryStatus } from '@suite-common/wallet-constants';
+import * as comparisonUtils from '@suite-common/suite-utils';
 
 import { ROUTER } from 'src/actions/suite/constants';
 import { selectDiscoveryForDevice, selectDevice } from 'src/reducers/suite/deviceReducer';
 import * as metadataActions from 'src/actions/suite/metadataActions';
-import * as comparisonUtils from 'src/utils/suite/comparisonUtils';
 import { getSelectedAccount } from 'src/utils/wallet/accountUtils';
 import { Action, Dispatch, GetState, AppState } from 'src/types/suite';
