+++ packages/suite/src/components/suite/modals/TorLoading.tsx
@@ -1,7 +1,8 @@
 import styled from 'styled-components';
 
+import { UserContextPayload } from '@suite-common/suite-types';
+
 import { TorLoader, Modal } from 'src/components/suite';
-import type { UserContextPayload } from 'src/actions/suite/modalActions';
 
 const SmallModal = styled(Modal)`
     width: 560px;
