+++ packages/suite/src/hooks/suite/useFilteredModal.ts
@@ -1,7 +1,8 @@
+import { UserContextPayload } from '@suite-common/suite-types';
+
 import { useSelector } from 'src/hooks/suite';
 import { MODAL } from 'src/actions/suite/constants';
 import type { State as ModalState } from 'src/reducers/suite/modalReducer';
-import type { UserContextPayload } from 'src/actions/suite/modalActions';
 
 type ModalContext = ModalState['context'];
 type UserModalType = UserContextPayload['type'];
