+++ packages/suite/src/hooks/suite/usePreferredModal.ts
@@ -1,6 +1,8 @@
+import { Route } from '@suite-common/suite-types';
+
 import { MODAL } from 'src/actions/suite/constants';
 import { useSelector, useDiscovery } from 'src/hooks/suite';
-import type { Route, ForegroundAppRoute } from 'src/types/suite';
+import type { ForegroundAppRoute } from 'src/types/suite';
 import { ModalAppParams } from 'src/utils/suite/router';
 
 const isForegroundApp = (route: Route): route is ForegroundAppRoute =>
