+++ packages/suite/src/types/suite/index.ts
@@ -7,6 +7,7 @@ import type { ObjectValues } from '@trezor/type-utils';
 import type { UiEvent, DeviceEvent, TransportEvent, BlockchainEvent } from '@trezor/connect';
 import { notificationsActions } from '@suite-common/toast-notifications';
 import { messageSystemActions } from '@suite-common/message-system';
+import type { Route } from '@suite-common/suite-types';
 
 import type { RouterAction } from 'src/actions/suite/routerActions';
 import type { AppState } from 'src/reducers/store';
@@ -24,7 +25,6 @@ import type { BackupAction } from 'src/actions/backup/backupActions';
 import type { RecoveryAction } from 'src/actions/recovery/recoveryActions';
 import type { SUITE } from 'src/actions/suite/constants';
 import type { GuideAction } from 'src/actions/suite/guideActions';
-import type { Route } from 'src/constants/suite/routes';
 import { deviceActions } from 'src/actions/suite/deviceActions';
 
 // reexport
