+++ suite-common/suite-config/src/index.ts
@@ -1,3 +1,5 @@
 export * from './features';
 export * from './fiat';
+export * from './tor';
+export * from './routes';
 export * from './settings';
