+++ suite-common/suite-config/src/tor.ts
@@ -1,0 +1,5 @@
+export enum RequestEnableTorResponse {
+    Continue = 'Continue',
+    Back = 'Back',
+    Skip = 'Skip',
+}
