+++ suite-common/suite-types/package.json
@@ -12,9 +12,13 @@
     "dependencies": {
         "@reduxjs/toolkit": "1.9.5",
         "@suite-common/metadata-types": "workspace:*",
+        "@suite-common/suite-config": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
+        "@suite-common/wallet-types": "workspace:*",
         "@trezor/connect": "workspace:*",
         "@trezor/env-utils": "workspace:*",
+        "@trezor/type-utils": "workspace:*",
+        "@trezor/utils": "workspace:*",
         "bignumber.js": "^9.1.1"
     },
     "devDependencies": {
