+++ suite-common/suite-types/src/index.ts
@@ -4,8 +4,10 @@ export * from './device';
 export * from './guide';
 export * from './firmware';
 export * from './sign';
+export * from './modal';
 export * from './github';
 export * from './messageSystem';
+export * from './route';
 
 export type Selector<TReturnValue> = (state: any) => TReturnValue;
 export type SuiteCompatibleAction<TPayload> = (
