+++ suite-common/suite-types/src/route.ts
@@ -1,0 +1,11 @@
+import { ArrayElement, ConstWithOptionalFields } from '@trezor/type-utils';
+import { routes } from '@suite-common/suite-config';
+
+type RouteKeys =
+    | keyof ArrayElement<typeof routes>
+    | 'params'
+    | 'exact'
+    | 'isForegroundApp'
+    | 'isFullscreenApp';
+
+export type Route = ArrayElement<ConstWithOptionalFields<typeof routes, RouteKeys>>;
