+++ suite-common/suite-utils/src/index.ts
@@ -4,3 +4,4 @@ export * from './features';
 export * from './build';
 export * from './antiFraud';
 export * from './resolveStaticPath';
+export * from './comparison';
